Gauss plume statistics

Calculates surface concentration of a Gaussian plumes resulting from a point source. This is done for a given time series of dispersion conditions as defined by wind speed, wind direction stability classes, and boundary layer height. The individual plumes are then averaged to provide the mean concentration during the given period.

average.gauss.plume.from.freq(
  xrng = c(-5000, 5000),
  yrng = c(-5000, 5000),
  dx = 50,
  dy = 50,
  freq,
  Q,
  h.e,
  C.ex = NULL
)

Arguments

xrng

Range of grid in x direction, given in m. Needs to be a two element vector.

yrng

Range of grid in y direction, given in m. Needs to be a two element vector.

dx

Resolution of grid in x direction, given in m.

dy

Resolution of gird in y direction, given in m.

freq

a data.frame containing the columns WS for wind speed in m/s, WD for wind direction in degree, 'stability' containing the stability class (A-F), 'h.m' the boundary layer height in m above ground, 'freq' the frequency of the conditions. If 'freq' is not contained in the data.frame, it is assumed to be 1 for all conditions.

Q

emission flux of the source. Given in kg/s.

h.e

effective emission height. Given in m above ground.

C.ex

Threshold concentration given in units ug/m3. If given an additional field counting the number of situations that exceed the given threshold is reported.

Value

Average concentrations for the area defined by xrng and yrng, given as a list with elements xrng, yrng, and zz (containing the concentrations in units ug/m3). If 'C.ex' was given and additional field 'nn.ex' will be returned giving the number of threshold exceedances per year.